refactor: the pull request minor comments
