+++ src/setupTests.ts
@@ -1,4 +1,3 @@
-import 'msw';
 import { beforeAll, afterAll, afterEach } from 'vitest';
 import { cleanup } from '@testing-library/react';
 import '@testing-library/jest-dom';
