--- src/store/gqlapi/graphQL/types.ts
@@ -1,4 +1,0 @@
-export interface GraphQLRequestBody {
-  query: string;
-  variables: string;
-}
